Tweaks

+++ setup/boot/init.d/locale.sh
@@ -1,18 +1,27 @@
-LOCALE_TZ="Europe/Brussels"
-LOCALE_KEYBOARD_LAYOUT="us"
-LOCALE_KEYBOARD_MODEL="pc105"
+#!/usr/bin/env bash
+# --------------------------------------------------------------
+set -euo pipefail # Stop running the script on first error...
+# --------------------------------------------------------------
+: "${LOCALE_TZ:=""}"
+: "${KEYBOARD_LAYOUT:=""}"
+: "${KEYBOARD_MODEL:=""}"
 
-# Set Timezone
-rm -f /etc/localtime
-echo "$LOCALE_TZ" >/etc/timezone
-dpkg-reconfigure -f noninteractive tzdata
+# Set Timezone (if specified)
+if [ ! -z "${LOCALE_TZ:-}" ]; then
+    rm -f /etc/localtime
+    echo "$LOCALE_TZ" >/etc/timezone
+    dpkg-reconfigure -f noninteractive tzdata
+fi
 
-# Set Keyboard layout
-cat >/etc/default/keyboard <<'KBEOF'
+# Change keyboard layout (if specified)
+if [ ! "${KEYBOARD_LAYOUT:-}" == "" ] && [ ! "${KEYBOARD_MODEL:-}" == "" ]; then
+    # Set Keyboard layout
+    cat >/etc/default/keyboard <<'KBEOF'
 XKBMODEL="$LOCALE_KEYBOARD_MODEL"
 XKBLAYOUT="$LOCALE_KEYBOARD_LAYOUT"
 XKBVARIANT=""
 XKBOPTIONS=""
 
 KBEOF
-dpkg-reconfigure -f noninteractive keyboard-configuration
+    dpkg-reconfigure -f noninteractive keyboard-configuration
+fi
